restructure into separate files
